Introduced caching timeout and enhanced statistics log

+++ urls/urls.py
@@ -17,31 +17,40 @@ class Urls:
     	"domain_fucus_only":True
     }
 
-    def __init__(self, seed, directory, cache = True):
+    def __init__(self, seed, directory, cache = True, cache_timeout = 100):
         self._path = "%s/urls"%directory
+        self._urls_path = self._path+"/urls.json"
         self._seed = seed
         self._cache = cache
-        self._url_trie = Trie(self._path+"/url_trie.pkl", cache=cache)
+        self._cache_timeout = cache_timeout
+        self._url_trie = Trie(self._path+"/url_trie.pkl", cache=cache, cache_timeout = cache_timeout)
         self._seed_domain = Urls.domain(self._seed)
 
         if not os.path.exists(self._path):
             os.makedirs(self._path)
 
     def _is_cached(self):
-        return self._cache and os.path.isfile(self._path)
+        return self._cache and os.path.isfile(self._urls_path)
 
     def _load_cache(self):
-        with open(self._path+"/urls.json") as f:
+        with open(self._urls_path) as f:
             data = json.load(f)
 
         self._unparsed= list(filter(lambda u: self.valid(u), data["unparsed"]))
 
     def _update_cache(self):
         if self._cache:
-            with open(self._path+"/urls.json", 'w') as f:
-                json.dump({
-                    "unparsed":self._working+self._unparsed
-                }, f)
+            if os.path.isfile(self._urls_path):
+                if os.path.getmtime(self._urls_path) > time.time() - self._cache_timeout:
+                    with open(self._urls_path, 'w') as f:
+                        json.dump({
+                            "unparsed":self._working+self._unparsed
+                        }, f)
+            else:
+                with open(self._urls_path, 'w') as f:
+                    json.dump({
+                        "unparsed":self._working+self._unparsed
+                    }, f)
 
     def set_validation_options(self, opt):
     	self._opt = {**self._opt, **opt}
@@ -88,6 +97,11 @@ class Urls:
 
     def domain(url):
         return '{uri.netloc}'.format(uri=urlparse(url))
+
+    def time_to_next_caching(self):
+        if self._cache:
+            return self._cache_timeout + os.path.getmtime(self._urls_path) - time.time()
+        return False
 
     def load(self):
         if self._is_cached():
